+++ packages/ast/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ts-linq/ast
 
+## 3.2.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @ts-linq/types@4.4.0
+
 ## 3.2.3
 
 ### Patch Changes